refactor(planning): split the 1,547-line engine along its seams
Fixes-program item 2. planning.engine.ts was the largest file in the codebase —
five concerns tangled in one class. Now four modules, each documenting its own
responsibility, with the engine as coordinator:

  planning/types.ts            (149)  shared types + PlanningDispositions (the
                                       Channel-A trait knobs as ONE mutable object
                                       read live by all modules — no copy drift)
  planning/plan.store.ts       (178)  canonical store, goal index (P4), terminal
                                       persist-once + retention GC (P5)
  planning/plan.frontier.ts    (86)   ready-set DAG + plan.prior projection —
                                       pure functions (planning-as-prior's heart)
  planning/plan.supervision.ts (~430) emergent tier, facet lifecycle, judgment
                                       vocabulary, directive dispatch; effects
                                       return through the narrow SupervisionHost
                                       interface (one-directional dependency)
  planning.engine.ts           (~640) engine shell: bus events, react loop,
                                       ingestion, execution, lifecycle, activity API

Mechanical extraction — bodies moved verbatim; only member access adapted
(this._plans.get → this._store.get etc.). Public import surface unchanged:
'#faculties/planning.engine' re-exports every previously-exported name.

Verification: typecheck ✓ · suite 892 pass / 1 skip / 0 fail (incl. the 8
planning behavioral test files) · replay-equivalence capstone ×2 green — the
byte-identity gate proves even the judgment prompt strings are unchanged.
dist rebuilt.
